chore(sdk-coin-avaxc): move account-lib unit tests
- moves unit tests from account-lib
- updates imports where necessary

Ticket: BG-48144
