+++ modules/sdk-coin-avaxc/src/index.ts
@@ -1,3 +1,4 @@
 export * from './avaxc';
 export * from './avaxcToken';
 export * from './tavaxc';
+export * from './lib';
