+++ modules/sdk-coin-avaxc/test/resources/avaxc.ts
