+++ modules/sdk-coin-avaxc/test/unit/getBuilder.ts
@@ -1,0 +1,6 @@
+import { TransactionBuilder } from '../../src';
+import { coins } from '@bitgo/statics';
+
+export const getBuilder = (coin: string): TransactionBuilder => {
+  return new TransactionBuilder(coins.get(coin));
+};
