+++ modules/sdk-coin-avaxc/test/unit/keyPair.ts
@@ -1,7 +1,7 @@
 import assert from 'assert';
 import should from 'should';
-import { KeyPair } from '../../../../src/coin/avaxc/keyPair';
-import { TEST_ACCOUNT } from '../../../resources/avaxc/avaxc';
+import { KeyPair } from '../../src';
+import { TEST_ACCOUNT } from '../resources/avaxc';
 
 describe('Avax Key Pair', () => {
   describe('should create a valid KeyPair', () => {
@@ -103,14 +103,14 @@ describe('Avax Key Pair', () => {
     it('from an invalid privateKey', () => {
       assert.throws(
         () => new KeyPair({ prv: '' }),
-        (e) => e.message === 'Unsupported private key',
+        (e) => e.message === 'Unsupported private key'
       );
     });
 
     it('from an invalid publicKey', () => {
       assert.throws(
         () => new KeyPair({ pub: '' }),
-        (e) => e.message.startsWith('Unsupported public key'),
+        (e) => e.message.startsWith('Unsupported public key')
       );
     });
 
@@ -118,7 +118,7 @@ describe('Avax Key Pair', () => {
       const undefinedBuffer = undefined as unknown as Buffer;
       assert.throws(
         () => new KeyPair({ seed: undefinedBuffer }),
-        (e) => e.message.startsWith('Invalid key pair options'),
+        (e) => e.message.startsWith('Invalid key pair options')
       );
     });
 
@@ -126,7 +126,7 @@ describe('Avax Key Pair', () => {
       const undefinedStr: string = undefined as unknown as string;
       assert.throws(
         () => new KeyPair({ prv: undefinedStr }),
-        (e) => e.message.startsWith('Invalid key pair options'),
+        (e) => e.message.startsWith('Invalid key pair options')
       );
     });
 
@@ -134,7 +134,7 @@ describe('Avax Key Pair', () => {
       const undefinedStr: string = undefined as unknown as string;
       assert.throws(
         () => new KeyPair({ pub: undefinedStr }),
-        (e) => e.message.startsWith('Invalid key pair options'),
+        (e) => e.message.startsWith('Invalid key pair options')
       );
     });
   });
